bugs resolved
