+++ src/Sections/Hostels/Hostels.css
@@ -125,3 +125,6 @@
     font-size: 1.4rem;
     font-weight: 600;
   }
+  .wwww{
+    border: 1px solid red;
+  }
